fixed compilation flags!!

+++ inc/writers.hpp
@@ -7,8 +7,9 @@
 #include <iostream>
 #include <sstream>
 #include <iomanip>
+#include <cmath>
 
 void writeToFile(std::ofstream&,  const Particle&, char); // To follow a single particle and its trajectory
 void writeDiagnosticsToFile(const Histogram, const int, const double, const std::string*, const int); // To write the Diagnostics to a numbered file
 void parseVector(const std::string&, double[3]); // To make a vector a string
-void printProgressBar(int, int, int); // Print progress bar during simulation
+void printProgressBar(double, double, int); // Print progress bar during simulation
